requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.3.0
pandas>=1.3.0
PyQt5>=5.15.0
albumentations>=1.1.0
Pillow>=8.3.0
scikit-learn>=1.0.0
pyyaml>=5.4.0
tqdm>=4.62.0